Updated lcd module to use ints. Added drv_usb.
Move all source files into labwiz folder.  Moved most includes
into labwiz folder.  Created system folder for headers required
by modules in the system.  Added include/system/ to include
path in project.

Deleted old and un-used source files

+++ trunk-labwiz/LabWizFirmware/include/labwiz/drv_lcd.h
@@ -31,13 +31,13 @@ void lcd_task( void *pvParameters );
 
 void lcd_blank(void);
 
-void lcd_set_pixel(uint8_t row, uint8_t col);
+void lcd_set_pixel(int row, int col);
 
-void lcd_clear_pixel(uint8_t row, uint8_t col);
+void lcd_clear_pixel(int row, int col);
 
-void lcd_print(char * st,uint8_t row,uint8_t col);
+void lcd_print(char * st,int row,int col);
 
-void lcd_line(uint8_t r0, uint8_t c0, uint8_t r1, uint8_t c1);
+void lcd_line(int r0, int c0, int r1, int c1);
 
 void lcd_backlight_toggle(void);
